+++ client/src/App.css
@@ -15,7 +15,7 @@
 
 .App-header {
   background-color: #282c34;
-  min-height: 100vh;
+  min-height: 20rem;
   display: flex;
   flex-direction: column;
   align-items: center;
