+++ client/src/App3.js
@@ -1,0 +1,13 @@
+import { useState } from 'react'
+import {
+    BrowserRouter as Router,
+    Routes,
+    Route,
+    Link
+} from 'react-router-dom';
+
+function App() {
+
+}
+
+export default App
