+++ client/src/component/loginS.js
@@ -1,0 +1,50 @@
+
+import React from 'react';
+import { useState } from 'react'
+
+function Login (){
+  const [email, setEmail] = useState('')
+  const [password, setPassword] = useState('')
+
+  async function loginUser(event) {
+    event.preventDefault()
+
+    const response = await fetch('http://localhost:5000/', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        email,
+        password,
+      }),
+    })
+
+    const data = await response.json()
+  }
+
+  return (
+    <div>
+      <h1>Login</h1>
+      <form onSubmit={loginUser}>
+        <input
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          type="email"
+          placeholder="Email"
+        />
+        <br />
+        <input
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          type="password"
+          placeholder="Password"
+        />
+        <br />
+        <input type="submit" value="Login" />
+      </form>
+    </div>
+  )
+}
+
+export default Login;
