+++ server/models/users.json
@@ -6,5 +6,21 @@
   "aeaeae@easdasaea.coasdasm": {
     "name": "aeae",
     "password": "asdasdaeaeae"
+  },
+  "teste@teste.com": {
+    "name": "Teste",
+    "password": "teste"
+  },
+  "teste3@teste.com": {
+    "name": "teadsad",
+    "password": "asdasdasd"
+  },
+  "arthur@email.com": {
+    "name": "Arthur",
+    "password": "senha"
+  },
+  "gomes@gomes.com": {
+    "name": "gomes",
+    "password": "gomes"
   }
 }
