+++ server/server.js
@@ -1,10 +1,10 @@
 const express = require('express');
+const cors = require('cors');
 var fs = require('fs');
 var users_data = fs.readFileSync('./models/users.json');
 var users = JSON.parse(users_data);
-const cors = require('cors');
 
-console.log(users);
+//console.log(users);
 
 const app = express();
 app.use(cors())
@@ -16,53 +16,7 @@ app.get('/api', (req, res) => {
   });
 });
 
-app.post('/api/registerUser', async (req, res) => {
-  var name = req.body.name;
-  var email = req.body.email;
-  var password = req.body.password;
-  users[email] = {name,password}
-  var data = JSON.stringify(users, null, 2);
-  fs.writeFile('./models/users.json', data, finished);
-
-  function finished(err) {
-    reply = {
-      name: name,
-      email: email,
-      status: 'success'
-    };
-    res.send(reply);
-  }
-})
-
-app.post('/api/loginUser', async (req, res) => {
-  var email = req.body.email;
-  var password = req.body.password;
-  var validated = false;
-
-  Object.entries(users).forEach((entry) => {
-    const [key, value] = entry;
-    if(email === key){
-      if(value['password'] === password){
-        validated = true;
-      }else{
-        validated = false;
-      }
-    }else{
-      validated = false;
-    }
-
-    if(validated){
-      reply = {
-        status: 'success'
-      };
-      res.send(reply);
-    }else{
-      reply = {
-        status: 'fail'
-      };
-      res.send(reply);
-    }
-  });
-})
+const userRoutes = require('./routes/users/users')
+app.use(userRoutes);
 
 app.listen(5000, () => console.log(`Listening on port ${5000}`));
